requirements.txt
flask==2.3.2
langchain==0.0.204
langchain_community==0.0.7
sqlalchemy-bigquery==1.12.0
google-cloud-bigquery==3.7.0
google-cloud-bigquery-storage==2.27.0
openai==0.27.0
gunicorn==20.1.0
